refactor(frontend): extract useGanttTaskActions hook from GanttChart
Bundle the Gantt chart's task/group navigation and edit actions — open plan,
open culture, add plan, open areas page, copy summary, delete plan, and the
double-click-to-plan wrapper — into a useGanttTaskActions hook. The page
passes navigate, plantingPlans, setPlantingPlans, setError, and t, and feeds
the returned callbacks into useGanttContextMenu and onTaskDoubleClick.
GanttChart drops the confirmAction, copyTextToClipboardSilently,
formatCultureDisplayLabel, and formatGanttDate imports it no longer uses
directly. No behavior change.
